+++ validator/sawtooth_validator/journal/block_validator.py
@@ -59,37 +59,6 @@ class ChainHeadUpdated(Exception):
     processing and restart processing with the new chain head.
     """
 
-class ForkResolutionError(Exception):
-    """
-    Indication that an error occured during fork resolution.
-    """
-
-class ForkResolutionResult:
-    def __init__(self, block):
-        self.block = block
-        self.chain_head = None
-        self.new_chain = []
-        self.current_chain = []
-        self.committed_batches = []
-        self.uncommitted_batches = []
-        # NOTE: The following are for all blocks validated in order to validate
-        # this block, i.e., all blocks on this block's fork
-        self.execution_results = []
-        self.transaction_count = 0
-
-    def __bool__(self):
-        return self.block.status == BlockStatus.Valid
-
-    def __str__(self):
-        keys = ("block", "valid", "chain_head", "new_chain", "current_chain",
-                "committed_batches", "uncommitted_batches",
-                "execution_results", "transaction_count")
-
-        out = "{"
-        for key in keys:
-            out += "%s: %s," % (key, self.__getattribute(key))
-        return out[:-1] + "}"
-
 
 # Need to disable this new pylint check until the function can be refactored
 # to return instead of raise StopIteration, which it does by calling next()
@@ -165,9 +134,6 @@ class BlockValidator(object):
 
         self._thread_pool = InstrumentedThreadPoolExecutor(1) \
             if thread_pool is None else thread_pool
-
-        self._moved_to_fork_count = COLLECTOR.counter(
-            'chain_head_moved_to_fork_count', instance=self)
 
         # Blocks that are currently being processed
         self._blocks_processing = ConcurrentSet()
@@ -314,6 +280,20 @@ class BlockValidator(object):
         # pylint: disable=broad-except
         try:
             try:
+                prev_block = self._block_cache[blkw.previous_block_id]
+            except KeyError:
+                prev_block = None
+            else:
+                if prev_block.status == BlockStatus.Invalid:
+                    raise BlockValidationFailure(
+                        "Block {} rejected due to invalid predecessor"
+                        " {}".format(blkw, prev_block))
+                elif prev_block.status == BlockStatus.Unknown:
+                    raise BlockValidationError(
+                        "Attempted to validate block {} before its predecessor"
+                        " {}".format(blkw, prev_block))
+
+            try:
                 prev_state_root = self._get_previous_block_state_root(blkw)
             except KeyError:
                 raise BlockValidationError(
@@ -323,11 +303,6 @@ class BlockValidator(object):
             if not self._validate_permissions(blkw, prev_state_root):
                 raise BlockValidationFailure(
                     'Block {} failed permission validation'.format(blkw))
-
-            try:
-                prev_block = self._block_cache[blkw.previous_block_id]
-            except KeyError:
-                prev_block = None
 
             consensus = self._load_consensus(prev_block)
             public_key = \
@@ -366,89 +341,6 @@ class BlockValidator(object):
                 "Unhandled exception BlockValidator.validate_block()")
             raise e
 
-    @staticmethod
-    def _compare_chain_height(head_a, head_b):
-        """Returns True if head_a is taller, False if head_b is taller, and
-        True if the heights are the same."""
-        return head_a.block_num - head_b.block_num >= 0
-
-    def _build_fork_diff_to_common_height(self, head_long, head_short):
-        """Returns a list of blocks on the longer chain since the greatest
-        common height between the two chains. Note that the chains may not
-        have the same block id at the greatest common height.
-
-        Args:
-            head_long (BlockWrapper)
-            head_short (BlockWrapper)
-
-        Returns:
-            (list of BlockWrapper) All blocks in the longer chain since the
-            last block in the shorter chain. Ordered newest to oldest.
-
-        Raises:
-            BlockValidationError
-                The block is missing a predecessor. Note that normally this
-                shouldn't happen because of the completer."""
-        fork_diff = []
-
-        last = head_short.block_num
-        blk = head_long
-
-        while blk.block_num > last:
-            if blk.previous_block_id == NULL_BLOCK_IDENTIFIER:
-                break
-
-            fork_diff.append(blk)
-            try:
-                blk = self._block_cache[blk.previous_block_id]
-            except KeyError:
-                raise ForkResolutionError(
-                    'Failed to build fork diff: block {} missing predecessor'
-                    .format(blk))
-
-        return blk, fork_diff
-
-    def _extend_fork_diff_to_common_ancestor(
-        self, new_blkw, cur_blkw, new_chain, cur_chain
-    ):
-        """ Finds a common ancestor of the two chains. new_blkw and cur_blkw
-        must be at the same height, or this will always fail.
-        """
-        while cur_blkw.identifier != new_blkw.identifier:
-            if (cur_blkw.previous_block_id == NULL_BLOCK_IDENTIFIER
-                    or new_blkw.previous_block_id == NULL_BLOCK_IDENTIFIER):
-                # We are at a genesis block and the blocks are not the same
-                for b in new_chain:
-                    b.status = BlockStatus.Invalid
-                raise ForkResolutionError(
-                    'Block {} rejected due to wrong genesis {}'.format(
-                        cur_blkw, new_blkw))
-
-            new_chain.append(new_blkw)
-            try:
-                new_blkw = self._block_cache[new_blkw.previous_block_id]
-            except KeyError:
-                raise ForkResolutionError(
-                    'Block {} rejected due to missing predecessor {}'.format(
-                        new_blkw, new_blkw.previous_block_id))
-
-            cur_chain.append(cur_blkw)
-            cur_blkw = self._block_cache[cur_blkw.previous_block_id]
-
-    def _compare_forks_consensus(self, chain_head, new_block):
-        """Ask the consensus module which fork to choose.
-        """
-        public_key = self._identity_signer.get_public_key().as_hex()
-        consensus = self._load_consensus(chain_head)
-        fork_resolver = consensus.ForkResolver(
-            block_cache=self._block_cache,
-            state_view_factory=self._state_view_factory,
-            data_dir=self._data_dir,
-            config_dir=self._config_dir,
-            validator_id=public_key)
-
-        return fork_resolver.compare_forks(chain_head, new_block)
-
     def _load_consensus(self, block):
         """Load the consensus module using the state as of the given block."""
         if block is not None:
@@ -458,24 +350,6 @@ class BlockValidator(object):
                     block,
                     self._state_view_factory))
         return ConsensusFactory.get_consensus_module('genesis')
-
-    @staticmethod
-    def _get_batch_commit_changes(new_chain, cur_chain):
-        """
-        Get all the batches that should be committed from the new chain and
-        all the batches that should be uncommitted from the current chain.
-        """
-        committed_batches = []
-        for blkw in new_chain:
-            for batch in blkw.batches:
-                committed_batches.append(batch)
-
-        uncommitted_batches = []
-        for blkw in cur_chain:
-            for batch in blkw.batches:
-                uncommitted_batches.append(batch)
-
-        return (committed_batches, uncommitted_batches)
 
     def submit_blocks_for_verification(self, blocks, callback):
         for block in blocks:
@@ -499,6 +373,24 @@ class BlockValidator(object):
                 self._add_block_to_pending(block)
                 continue
 
+            try:
+                prev_block = self._block_cache[block.previous_block_id]
+            except KeyError:
+                LOGGER.error(
+                    "Block %s submitted for processing but predecessor %s is"
+                    " missing. Adding to pending.",
+                    block,
+                    block.previous_block_id)
+                self._add_block_to_pending(block)
+                continue
+            else:
+                if prev_block.status == BlockStatus.Unknown:
+                    LOGGER.warning(
+                        "Block %s submitted for processing but predecessor %s"
+                        " has not been validated and is not pending. Adding to"
+                        " pending.", block, prev_block)
+                    self._add_block_to_pending(block)
+
             LOGGER.debug(
                 "Adding block %s for processing", block.identifier)
 
@@ -512,8 +404,7 @@ class BlockValidator(object):
 
     def _wrap_callback(self, block, callback):
         # Internal cleanup after verification
-        def wrapper(commit_new_block, result):
-            block = result.block
+        def wrapper(block):
             LOGGER.debug("Removing block from processing %s", block.identifier)
             try:
                 self._blocks_processing.remove(block.identifier)
@@ -572,7 +463,7 @@ class BlockValidator(object):
                         self._blocks_pending_descendants.pop(
                             block.identifier, []))
 
-            callback(commit_new_block, result)
+            callback(block)
 
         return wrapper
 
@@ -594,105 +485,40 @@ class BlockValidator(object):
         be the new head block. Returns the results to the ChainController
         so that the change over can be made if necessary.
         """
-        try:
-            result = ForkResolutionResult(block)
-            LOGGER.info("Starting block validation of : %s", block)
-
-            # Get the current chain_head and store it in the result
+        while True:
             chain_head = self._block_cache.block_store.chain_head
-            result.chain_head = chain_head
-
             try:
                 self.validate_block(block)
+                LOGGER.info(
+                    'Block %s passed validation', block)
             except BlockValidationFailure as err:
                 LOGGER.warning(
-                    'Block %s failed validation: %s',
-                    block, err)
-                block.status = BlockStatus.Invalid
-                callback(False, result)
-                return
+                    'Block %s failed validation: %s', block, err)
             except BlockValidationError as err:
                 LOGGER.error(
-                    'Encountered an error while validating %s: %s', blk, err)
-                callback(False, result)
-                return
+                    'Encountered an error while validating %s: %s', block, err)
+            except Exception:  # pylint: disable=broad-except
+                LOGGER.exception(
+                    "Block validation failed with unexpected error: %s", block)
 
-            # Create new local variables for current and new block, since
-            # these variables get modified later
-            current_block = chain_head
-            new_block = block
-
-            try:
-                # Get all the blocks since the greatest common height from the
-                # longer chain.
-                if self._compare_chain_height(current_block, new_block):
-                    current_block, result.current_chain =\
-                        self._build_fork_diff_to_common_height(
-                            current_block, new_block)
+            # The validity of blocks depends partially on whether or not there
+            # are any duplicate transactions or batches in the block. This can
+            # only be checked accurately if the block store does not update
+            # during validation. The current practice is the assume this will
+            # not happen and, if it does, to reprocess the validation. This
+            # has been experimentally proven to be more performant than locking
+            # the chain head and block store around duplicate checking.
+            if chain_head is None:
+                break
+            else:
+                current_chain_head = self._block_cache.block_store.chain_head
+                if chain_head.identifier == current_chain_head.identifier:
+                    break
                 else:
-                    new_block, result.new_chain =\
-                        self._build_fork_diff_to_common_height(
-                            new_block, current_block)
+                    LOGGER.warning(
+                        "Chain head updated from %s to %s while validating"
+                        " block %s. Reprocessing validation.",
+                        chain_head, current_chain_head, block)
+                    block.status = BlockStatus.Unknown
 
-                # Add blocks to the two chains until a common ancestor is found
-                # or raise an exception if no common ancestor is found
-                self._extend_fork_diff_to_common_ancestor(
-                    new_block, current_block,
-                    result.new_chain, result.current_chain)
-            except ForkResolutionError as err:
-                LOGGER.error(
-                    'Encountered an error while resolving a fork with head %s:'
-                    ' %s', block, err)
-                callback(False, result)
-                return
-
-            for blk in reversed(result.new_chain):
-                result.transaction_count += block.num_transactions
-
-            # Ask consensus if the new chain should be committed
-            LOGGER.info(
-                "Comparing current chain head '%s' against new block '%s'",
-                chain_head, new_block)
-            for i in range(max(
-                len(result.new_chain), len(result.current_chain)
-            )):
-                cur = new = num = "-"
-                if i < len(result.current_chain):
-                    cur = result.current_chain[i].header_signature[:8]
-                    num = result.current_chain[i].block_num
-                if i < len(result.new_chain):
-                    new = result.new_chain[i].header_signature[:8]
-                    num = result.new_chain[i].block_num
-                LOGGER.info(
-                    "Fork comparison at height %s is between %s and %s",
-                    num, cur, new)
-
-            commit_new_chain = self._compare_forks_consensus(chain_head, block)
-
-            # If committing the new chain, get the list of committed batches
-            # from the current chain that need to be uncommitted and the list
-            # of uncommitted batches from the new chain that need to be
-            # committed.
-            if commit_new_chain:
-                commit, uncommit =\
-                    self._get_batch_commit_changes(
-                        result.new_chain, result.current_chain)
-                result.committed_batches = commit
-                result.uncommitted_batches = uncommit
-
-                if result.new_chain[0].previous_block_id \
-                        != chain_head.identifier:
-                    self._moved_to_fork_count.inc()
-
-            # Pass the results to the callback function
-            callback(commit_new_chain, result)
-            LOGGER.info("Finished block validation of: %s", block)
-
-        except ChainHeadUpdated:
-            callback(False, result)
-            return
-        except Exception:  # pylint: disable=broad-except
-            LOGGER.exception(
-                "Block validation failed with unexpected error: %s", block)
-            # callback to clean up the block out of the processing list.
-            callback(False, result)
+        callback(block)
