+++ validator/sawtooth_validator/journal/chain.py
@@ -20,8 +20,12 @@ import queue
 from threading import RLock
 
 from sawtooth_validator.concurrent.thread import InstrumentedThread
+from sawtooth_validator.journal.block_wrapper import BlockStatus
+from sawtooth_validator.journal.block_wrapper import BlockWrapper
 from sawtooth_validator.journal.block_wrapper import NULL_BLOCK_IDENTIFIER
 from sawtooth_validator.journal.block_validator import BlockValidationFailure
+from sawtooth_validator.journal.consensus.consensus_factory import \
+    ConsensusFactory
 from sawtooth_validator.protobuf.transaction_receipt_pb2 import \
     TransactionReceipt
 from sawtooth_validator import metrics
@@ -29,6 +33,39 @@ from sawtooth_validator import metrics
 
 LOGGER = logging.getLogger(__name__)
 COLLECTOR = metrics.get_collector(__name__)
+
+
+class ForkResolutionError(Exception):
+    """
+    Indication that an error occured during fork resolution.
+    """
+
+
+class ForkResolutionResult:
+    def __init__(self, block):
+        self.block = block
+        self.chain_head = None
+        self.new_chain = []
+        self.current_chain = []
+        self.committed_batches = []
+        self.uncommitted_batches = []
+        # NOTE: The following are for all blocks validated in order to validate
+        # this block, i.e., all blocks on this block's fork
+        self.execution_results = []
+        self.transaction_count = 0
+
+    def __bool__(self):
+        return self.block.status == BlockStatus.Valid
+
+    def __str__(self):
+        keys = ("block", "valid", "chain_head", "new_chain", "current_chain",
+                "committed_batches", "uncommitted_batches",
+                "execution_results", "transaction_count")
+
+        out = "{"
+        for key in keys:
+            out += "%s: %s," % (key, self.__getattribute(key))
+        return out[:-1] + "}"
 
 
 class ChainObserver(object, metaclass=ABCMeta):
@@ -84,6 +121,7 @@ class ChainController(object):
                  chain_head_lock,
                  on_chain_updated,
                  chain_id_manager,
+                 identity_signer,
                  data_dir,
                  config_dir,
                  chain_observers):
@@ -106,6 +144,7 @@ class ChainController(object):
             on_chain_updated: The callback to call to notify the rest of the
                  system the head block in the chain has been changed.
             chain_id_manager: The ChainIdManager instance.
+            identity_signer: A cryptographic signer for signing blocks.
             data_dir: path to location where persistent data for the
                 consensus module can be stored.
             config_dir: path to location where config data for the
@@ -121,6 +160,7 @@ class ChainController(object):
         self._block_store = block_cache.block_store
         self._state_view_factory = state_view_factory
         self._notify_on_chain_updated = on_chain_updated
+        self._identity_signer = identity_signer
         self._data_dir = data_dir
         self._config_dir = config_dir
 
@@ -139,6 +179,9 @@ class ChainController(object):
         self._block_num_gauge = COLLECTOR.gauge('block_num', instance=self)
         self._blocks_considered_count = COLLECTOR.counter(
             'blocks_considered_count', instance=self)
+
+        self._moved_to_fork_count = COLLECTOR.counter(
+            'chain_head_moved_to_fork_count', instance=self)
 
         self._block_queue = queue.Queue()
         self._chain_thread = None
@@ -192,7 +235,7 @@ class ChainController(object):
         self._block_validator.submit_blocks_for_verification(
             blocks, self.on_block_validated)
 
-    def on_block_validated(self, commit_new_block, result):
+    def on_block_validated(self, block):
         """Message back from the block validator, that the validation is
         complete
         Args:
@@ -203,25 +246,16 @@ class ChainController(object):
             None
         """
         try:
+            if block.status != BlockStatus.Valid:
+                return
+
             with self._lock:
+                commit_new_block, result = self._resolve_fork(block)
                 self._blocks_considered_count.inc()
                 new_block = result.block
 
-                # if the head has changed, since we started the work.
-                if result.chain_head.identifier !=\
-                        self._chain_head.identifier:
-                    LOGGER.info(
-                        'Chain head updated from %s to %s while processing '
-                        'block: %s',
-                        result.chain_head,
-                        self._chain_head,
-                        new_block)
-
-                    LOGGER.debug('Verify block again: %s ', new_block)
-                    self._submit_blocks_for_verification([new_block])
-
                 # If the head is to be updated to the new block.
-                elif commit_new_block:
+                if commit_new_block:
                     with self._chain_head_lock:
                         self._chain_head = new_block
 
@@ -257,11 +291,11 @@ class ChainController(object):
                                              batch.header_signature,
                                              self.__class__.__name__)
 
-                    for block in reversed(result.new_chain):
-                        receipts = self._make_receipts(block.execution_results)
+                    for blk in reversed(result.new_chain):
+                        receipts = self._make_receipts(blk.execution_results)
                         # Update all chain observers
                         for observer in self._chain_observers:
-                            observer.chain_update(block, receipts)
+                            observer.chain_update(blk, receipts)
 
                 # The block is otherwise valid, but we have determined we
                 # don't want it as the chain head.
@@ -272,6 +306,194 @@ class ChainController(object):
         except Exception:
             LOGGER.exception(
                 "Unhandled exception in ChainController.on_block_validated()")
+
+    def _resolve_fork(self, block):
+        result = ForkResolutionResult(block)
+        LOGGER.info("Starting fork resolution of : %s", block)
+
+        # Get the current chain_head and store it in the result
+        chain_head = self._block_cache.block_store.chain_head
+        result.chain_head = chain_head
+
+        # Create new local variables for current and new block, since
+        # these variables get modified later
+        current_block = chain_head
+        new_block = block
+
+        try:
+            # Get all the blocks since the greatest common height from the
+            # longer chain.
+            if self._compare_chain_height(current_block, new_block):
+                current_block, result.current_chain =\
+                    self._build_fork_diff_to_common_height(
+                        current_block, new_block)
+            else:
+                new_block, result.new_chain =\
+                    self._build_fork_diff_to_common_height(
+                        new_block, current_block)
+
+            # Add blocks to the two chains until a common ancestor is found
+            # or raise an exception if no common ancestor is found
+            self._extend_fork_diff_to_common_ancestor(
+                new_block, current_block,
+                result.new_chain, result.current_chain)
+        except ForkResolutionError as err:
+            LOGGER.error(
+                'Encountered an error while resolving a fork with head %s:'
+                ' %s', block, err)
+            return False, result
+
+        for blk in reversed(result.new_chain):
+            result.transaction_count += blk.num_transactions
+
+        # Ask consensus if the new chain should be committed
+        LOGGER.info(
+            "Comparing current chain head '%s' against new block '%s'",
+            chain_head, new_block)
+        for i in range(max(
+            len(result.new_chain), len(result.current_chain)
+        )):
+            cur = new = num = "-"
+            if i < len(result.current_chain):
+                cur = result.current_chain[i].header_signature[:8]
+                num = result.current_chain[i].block_num
+            if i < len(result.new_chain):
+                new = result.new_chain[i].header_signature[:8]
+                num = result.new_chain[i].block_num
+            LOGGER.info(
+                "Fork comparison at height %s is between %s and %s",
+                num, cur, new)
+
+        commit_new_chain = self._compare_forks_consensus(chain_head, block)
+
+        # If committing the new chain, get the list of committed batches
+        # from the current chain that need to be uncommitted and the list
+        # of uncommitted batches from the new chain that need to be
+        # committed.
+        if commit_new_chain:
+            commit, uncommit =\
+                self._get_batch_commit_changes(
+                    result.new_chain, result.current_chain)
+            result.committed_batches = commit
+            result.uncommitted_batches = uncommit
+
+            if result.new_chain[0].previous_block_id \
+                    != chain_head.identifier:
+                self._moved_to_fork_count.inc()
+
+        LOGGER.info("Finished fork resolution of: %s", block)
+        return commit_new_chain, result
+
+    @staticmethod
+    def _compare_chain_height(head_a, head_b):
+        """Returns True if head_a is taller, False if head_b is taller, and
+        True if the heights are the same."""
+        return head_a.block_num - head_b.block_num >= 0
+
+    def _build_fork_diff_to_common_height(self, head_long, head_short):
+        """Returns a list of blocks on the longer chain since the greatest
+        common height between the two chains. Note that the chains may not
+        have the same block id at the greatest common height.
+
+        Args:
+            head_long (BlockWrapper)
+            head_short (BlockWrapper)
+
+        Returns:
+            (list of BlockWrapper) All blocks in the longer chain since the
+            last block in the shorter chain. Ordered newest to oldest.
+
+        Raises:
+            BlockValidationError
+                The block is missing a predecessor. Note that normally this
+                shouldn't happen because of the completer."""
+        fork_diff = []
+
+        last = head_short.block_num
+        blk = head_long
+
+        while blk.block_num > last:
+            if blk.previous_block_id == NULL_BLOCK_IDENTIFIER:
+                break
+
+            fork_diff.append(blk)
+            try:
+                blk = self._block_cache[blk.previous_block_id]
+            except KeyError:
+                raise ForkResolutionError(
+                    'Failed to build fork diff: block {} missing predecessor'
+                    .format(blk))
+
+        return blk, fork_diff
+
+    def _extend_fork_diff_to_common_ancestor(
+        self, new_blkw, cur_blkw, new_chain, cur_chain
+    ):
+        """ Finds a common ancestor of the two chains. new_blkw and cur_blkw
+        must be at the same height, or this will always fail.
+        """
+        while cur_blkw.identifier != new_blkw.identifier:
+            if (cur_blkw.previous_block_id == NULL_BLOCK_IDENTIFIER
+                    or new_blkw.previous_block_id == NULL_BLOCK_IDENTIFIER):
+                # We are at a genesis block and the blocks are not the same
+                for b in new_chain:
+                    b.status = BlockStatus.Invalid
+                raise ForkResolutionError(
+                    'Block {} rejected due to wrong genesis {}'.format(
+                        cur_blkw, new_blkw))
+
+            new_chain.append(new_blkw)
+            try:
+                new_blkw = self._block_cache[new_blkw.previous_block_id]
+            except KeyError:
+                raise ForkResolutionError(
+                    'Block {} rejected due to missing predecessor {}'.format(
+                        new_blkw, new_blkw.previous_block_id))
+
+            cur_chain.append(cur_blkw)
+            cur_blkw = self._block_cache[cur_blkw.previous_block_id]
+
+    def _compare_forks_consensus(self, chain_head, new_block):
+        """Ask the consensus module which fork to choose.
+        """
+        public_key = self._identity_signer.get_public_key().as_hex()
+        consensus = self._load_consensus(chain_head)
+        fork_resolver = consensus.ForkResolver(
+            block_cache=self._block_cache,
+            state_view_factory=self._state_view_factory,
+            data_dir=self._data_dir,
+            config_dir=self._config_dir,
+            validator_id=public_key)
+
+        return fork_resolver.compare_forks(chain_head, new_block)
+
+    def _load_consensus(self, block):
+        """Load the consensus module using the state as of the given block."""
+        if block is not None:
+            return ConsensusFactory.get_configured_consensus_module(
+                block.header_signature,
+                BlockWrapper.state_view_for_block(
+                    block,
+                    self._state_view_factory))
+        return ConsensusFactory.get_consensus_module('genesis')
+
+    @staticmethod
+    def _get_batch_commit_changes(new_chain, cur_chain):
+        """
+        Get all the batches that should be committed from the new chain and
+        all the batches that should be uncommitted from the current chain.
+        """
+        committed_batches = []
+        for blkw in new_chain:
+            for batch in blkw.batches:
+                committed_batches.append(batch)
+
+        uncommitted_batches = []
+        for blkw in cur_chain:
+            for batch in blkw.batches:
+                uncommitted_batches.append(batch)
+
+        return (committed_batches, uncommitted_batches)
 
     def on_block_received(self, block):
         try:
