+++ validator/sawtooth_validator/server/core.py
@@ -302,6 +302,7 @@ class Validator(object):
             chain_head_lock=block_publisher.chain_head_lock,
             on_chain_updated=block_publisher.on_chain_updated,
             chain_id_manager=chain_id_manager,
+            identity_signer=identity_signer,
             data_dir=data_dir,
             config_dir=config_dir,
             chain_observers=[
